Changes according to discussion
